Adding exceptions and refactoring
