+++ sdc/crypto/encrypter.py
@@ -1,4 +1,5 @@
-from sdc.crypto.token_helper import JWTHelper, JWEHelper
+from sdc.crypto.jwe_helper import JWEHelper
+from sdc.crypto.jwt_helper import JWTHelper
 
 
 def encrypt(json, key_store, key_purpose):
