+++ sdc/crypto/exceptions.py
@@ -1,4 +1,8 @@
 
+class CryptoError(Exception):
+    pass
+
+
 class InvalidTokenException(Exception):
     def __init__(self, value="This token is invalid"):
         self.value = value
