+++ sdc/crypto/scripts/generate_secrets.py
@@ -9,20 +9,19 @@ from cryptography.hazmat.primitives.serialization import load_pem_private_key, E
 import yaml
 from yaml.representer import SafeRepresenter
 
-
 '''
   This script will generate a secrets.yml file with key information extracted from the provided folder. It assumes
   keys are in the format:
-  
+
   <platform>-<service>-<purpose>-<key_use>-<key_type>-<version>.pem
 
-    platform | Platform the keys are used on e.g. sdc 
-    service  | Service the key originates from (holder of the private key) e.g. rrm, sdx or eq 
+    platform | Platform the keys are used on e.g. sdc
+    service  | Service the key originates from (holder of the private key) e.g. rrm, sdx or eq
     purpose  | Purpose of the key in the service e.g. authentication or submission
     key_use  | encryption or signing
     key_type | public or private
     version  | Version identifier e.g. v1
-    
+
   e.g. sdc-rrm-authentication-encryption-public-v1.pem
 
   This script generates a yml file in the following format.
@@ -59,11 +58,14 @@ def change_style(style, representer):
     """
     This function is used to format the key value as a multi-line string maintaining the line breaks
     """
+
     def new_representer(dumper, data):
         scalar = representer(dumper, data)
         scalar.style = style
         return scalar
+
     return new_representer
+
 
 represent_literal_unicode = change_style('|', SafeRepresenter.represent_str)
 yaml.add_representer(LiteralUnicode, represent_literal_unicode)
@@ -77,7 +79,8 @@ def get_file_contents(folder, filename, trim=False):
     return data
 
 
-def _generate_kid_from_key_and_add_to_dict(keys, key_type, platform, service, purpose, key_use, version, public_key, private_key=None, kid_override=None):
+def _generate_kid_from_key_and_add_to_dict(keys, key_type, platform, service, purpose, key_use, version, public_key,
+                                           private_key=None, kid_override=None):
     if not kid_override:
         hash_object = hashlib.sha1(public_key.encode())
         kid = hash_object.hexdigest()
@@ -95,7 +98,8 @@ def _generate_kid_from_key_and_add_to_dict(keys, key_type, platform, service, pu
     keys[kid_override if kid_override else kid] = key
 
 
-def add_public_key_to_dict(keys, platform, service, purpose, key_use, version, public_key, keys_folder, kid_override=None):
+def add_public_key_to_dict(keys, platform, service, purpose, key_use, version, public_key, keys_folder,
+                           kid_override=None):
     '''
     Loads a public key from the file system and adds it to a dict of keys
     :param keys: A dict of keys
@@ -111,10 +115,12 @@ def add_public_key_to_dict(keys, platform, service, purpose, key_use, version, p
     '''
     public_key_data = get_file_contents(keys_folder, public_key)
 
-    _generate_kid_from_key_and_add_to_dict(keys, "public", platform, service, purpose, key_use, version, public_key_data, kid_override=kid_override)
+    _generate_kid_from_key_and_add_to_dict(keys, "public", platform, service, purpose, key_use, version,
+                                           public_key_data, kid_override=kid_override)
 
 
-def add_private_key_to_dict(keys, platform, service, purpose, key_use, version, private_key, keys_folder, kid_override=None):
+def add_private_key_to_dict(keys, platform, service, purpose, key_use, version, private_key, keys_folder,
+                            kid_override=None):
     '''
     Loads a private key from the file system and adds it to a dict of keys
     :param keys: A dict of keys
@@ -136,16 +142,16 @@ def add_private_key_to_dict(keys, platform, service, purpose, key_use, version, 
 
     pub_bytes = pub_key.public_bytes(Encoding.PEM, PublicFormat.SubjectPublicKeyInfo)
 
-    _generate_kid_from_key_and_add_to_dict(keys, "private", platform, service, purpose, key_use, version, pub_bytes.decode(), private_key_data, kid_override=kid_override)
+    _generate_kid_from_key_and_add_to_dict(keys, "private", platform, service, purpose, key_use, version,
+                                           pub_bytes.decode(), private_key_data, kid_override=kid_override)
 
 
-def generate_secrets_file(keys,):
+def generate_secrets_file(keys, ):
     with open('secrets.yml', 'w') as f:
         yaml.dump({"keys": keys}, f, default_flow_style=False)
 
 
 def generate_keys(keys_folder):
-
     key_files = [f for f in os.listdir(keys_folder) if os.path.isfile(os.path.join(keys_folder, f))]
 
     keys = {}
@@ -175,5 +181,4 @@ if __name__ == "__main__":
 
     args = parser.parse_args()
 
-    keys_folder = args.folder
-    generate_keys(keys_folder)
+    generate_keys(args.folder)
